+++ smallvil/src/winit.rs
@@ -115,7 +115,7 @@ pub fn winit_dispatch(
     let damage = Rectangle::from_loc_and_size((0, 0), size);
 
     backend.bind().ok().and_then(|_| {
-        smithay::desktop::space::render_output::<_, WaylandSurfaceRenderElement, _>(
+        smithay::desktop::space::render_output::<_, WaylandSurfaceRenderElement, _, _>(
             output,
             backend.renderer(),
             0,
@@ -123,7 +123,7 @@ pub fn winit_dispatch(
             &[],
             damage_tracked_renderer,
             [0.1, 0.1, 0.1, 1.0],
-            log,
+            log.clone(),
         )
         .unwrap()
     });
