+++ src/backend/renderer/element/surface.rs
@@ -188,7 +188,7 @@
 //!
 //!     // Render the element(s)
 //!     damage_tracked_renderer
-//!         .render_output(&mut renderer, 0, &*render_elements, [0.8, 0.8, 0.9, 1.0], &log)
+//!         .render_output(&mut renderer, 0, &*render_elements, [0.8, 0.8, 0.9, 1.0], log.clone())
 //!         .expect("failed to render output");
 //! }
 //! ```
