+++ src/backend/renderer/element/tests.rs
@@ -1,0 +1,277 @@
+use std::marker::PhantomData;
+
+use crate::{
+    backend::renderer::{gles2::Gles2Renderer, ImportDma, ImportMem, Renderer, Texture},
+    utils::{Buffer, Physical, Point, Rectangle, Scale},
+};
+
+use super::{CommitCounter, Id, RenderElement, Wrap};
+
+render_elements! {
+    ImportMemTest<R> where R: ImportMem;
+    Memory=ImportMemRenderElement,
+}
+
+render_elements! {
+    ImportMemTest2<=Gles2Renderer>;
+    Memory=ImportMemRenderElement,
+}
+
+render_elements! {
+    ImportMemTest3<'a, R, T> where R: ImportMem, T: Texture;
+    Memory=ImportMemRenderElement,
+    Custom=&'a T,
+}
+
+render_elements! {
+    ImportMemTest4<R> where R: ImportMem + ImportDma;
+    Memory=ImportMemRenderElement,
+}
+
+render_elements! {
+    ImportMemTest5<'a, R, T> where R: ImportMem + ImportDma, T: Texture;
+    Memory=ImportMemRenderElement,
+    Custom=&'a T,
+}
+
+render_elements! {
+    TextureIdTest<R> where R: ImportMem, <R as Renderer>::TextureId: Clone;
+    Memory=ImportMemRenderElement,
+}
+
+render_elements! {
+    TextureIdTest1<R> where R: ImportMem, <R as Renderer>::TextureId: 'static;
+    Memory=ImportMemRenderElement,
+}
+
+render_elements! {
+    TextureIdTest2<'a, R, C> where R: ImportMem, <R as Renderer>::TextureId: 'a;
+    Memory=ImportMemRenderElement,
+    Custom=&'a C,
+}
+
+render_elements! {
+    TextureIdTest3<'a, R, C> where R: ImportMem, <R as Renderer>::TextureId: Clone + 'a;
+    Memory=ImportMemRenderElement,
+    Custom=&'a C,
+}
+
+render_elements! {
+    Test<='a, Gles2Renderer>;
+    Surface=TestRenderElement<'a, Gles2Renderer>
+}
+
+render_elements! {
+    Test2<=Gles2Renderer>;
+    Surface=TestRenderElement2<Gles2Renderer>
+}
+
+render_elements! {
+    Test3<='a, Gles2Renderer, C>;
+    Surface=TestRenderElement<'a, Gles2Renderer>,
+    Custom=&'a C,
+}
+
+render_elements! {
+    Test4<=Gles2Renderer, C>;
+    Surface=TestRenderElement2<Gles2Renderer>,
+    Custom=C
+}
+
+render_elements! {
+    TestG<'a, R>;
+    Surface=TestRenderElement<'a, R>
+}
+
+render_elements! {
+    TestG2<R>;
+    Surface=TestRenderElement2<R>
+}
+
+render_elements! {
+    TestG3<'a, R, C>;
+    Surface=TestRenderElement<'a, R>,
+    Custom=&'a C,
+}
+
+render_elements! {
+    TestG4<R, C>;
+    Surface=TestRenderElement2<R>,
+    Custom=Wrap<C>
+}
+
+render_elements! {
+    TestG5;
+    What=Empty,
+}
+
+render_elements! {
+    TestG6<'a, R, C>;
+    Surface=TestRenderElement<'a, R>,
+    Custom=&'a C,
+    Custom2=Wrap<C>,
+}
+
+render_elements! {
+    TestG7<'a, R, C1, C2>;
+    Surface=TestRenderElement<'a, R>,
+    Custom=Wrap<C1>,
+    Custom2=Wrap<C2>,
+}
+
+struct ImportMemRenderElement {}
+
+impl<R> RenderElement<R> for ImportMemRenderElement
+where
+    R: Renderer + ImportMem,
+{
+    fn id(&self) -> &Id {
+        todo!()
+    }
+
+    fn current_commit(&self) -> CommitCounter {
+        todo!()
+    }
+
+    fn geometry(&self, _scale: Scale<f64>) -> Rectangle<i32, Physical> {
+        todo!()
+    }
+
+    fn draw(
+        &self,
+        _renderer: &mut R,
+        _frame: &mut <R as Renderer>::Frame,
+        _location: Point<i32, Physical>,
+        _scale: Scale<f64>,
+        _damage: &[Rectangle<i32, Physical>],
+        _log: &slog::Logger,
+    ) -> Result<(), <R as Renderer>::Error> {
+        todo!()
+    }
+
+    fn src(&self) -> Rectangle<f64, Buffer> {
+        todo!()
+    }
+}
+
+impl<R> RenderElement<R> for Empty
+where
+    R: Renderer,
+{
+    fn id(&self) -> &Id {
+        todo!()
+    }
+
+    fn current_commit(&self) -> CommitCounter {
+        todo!()
+    }
+
+    fn location(&self, _scale: Scale<f64>) -> Point<i32, Physical> {
+        todo!()
+    }
+
+    fn src(&self) -> Rectangle<f64, Buffer> {
+        todo!()
+    }
+
+    fn geometry(&self, _scale: Scale<f64>) -> Rectangle<i32, Physical> {
+        todo!()
+    }
+
+    fn draw(
+        &self,
+        _renderer: &mut R,
+        _frame: &mut <R as Renderer>::Frame,
+        _location: Point<i32, Physical>,
+        _scale: Scale<f64>,
+        _damage: &[Rectangle<i32, Physical>],
+        _log: &slog::Logger,
+    ) -> Result<(), <R as Renderer>::Error> {
+        todo!()
+    }
+}
+
+struct Empty;
+
+struct TestRenderElement2<R> {
+    _phantom: PhantomData<R>,
+}
+
+impl<R> RenderElement<R> for TestRenderElement2<R>
+where
+    R: Renderer,
+{
+    fn id(&self) -> &Id {
+        todo!()
+    }
+
+    fn current_commit(&self) -> CommitCounter {
+        todo!()
+    }
+
+    fn location(&self, _scale: Scale<f64>) -> Point<i32, Physical> {
+        todo!()
+    }
+
+    fn src(&self) -> Rectangle<f64, Buffer> {
+        todo!()
+    }
+
+    fn geometry(&self, _scale: Scale<f64>) -> Rectangle<i32, Physical> {
+        todo!()
+    }
+
+    fn draw(
+        &self,
+        _renderer: &mut R,
+        _frame: &mut <R as Renderer>::Frame,
+        _location: Point<i32, Physical>,
+        _scale: Scale<f64>,
+        _damage: &[Rectangle<i32, Physical>],
+        _log: &slog::Logger,
+    ) -> Result<(), <R as Renderer>::Error> {
+        todo!()
+    }
+}
+
+struct TestRenderElement<'a, R> {
+    _test: &'a usize,
+    _phantom: PhantomData<R>,
+}
+
+impl<'a, R> RenderElement<R> for TestRenderElement<'a, R>
+where
+    R: Renderer,
+{
+    fn id(&self) -> &Id {
+        todo!()
+    }
+
+    fn current_commit(&self) -> CommitCounter {
+        todo!()
+    }
+
+    fn location(&self, _scale: Scale<f64>) -> Point<i32, Physical> {
+        todo!()
+    }
+
+    fn src(&self) -> Rectangle<f64, Buffer> {
+        todo!()
+    }
+
+    fn geometry(&self, _scale: Scale<f64>) -> Rectangle<i32, Physical> {
+        todo!()
+    }
+
+    fn draw(
+        &self,
+        _renderer: &mut R,
+        _frame: &mut <R as Renderer>::Frame,
+        _location: Point<i32, Physical>,
+        _scale: Scale<f64>,
+        _damage: &[Rectangle<i32, Physical>],
+        _log: &slog::Logger,
+    ) -> Result<(), <R as Renderer>::Error> {
+        todo!()
+    }
+}
